test: move handler for testing
