+++ apps/vs-agent/tests/vtFlow.test.ts
@@ -1,16 +1,17 @@
 import { DidCommConnectionRecord, DidCommHandshakeProtocol } from '@credo-ts/didcomm'
 import { VtFlowRole, VtFlowState, VtFlowVariant } from '@verana-labs/credo-ts-didcomm-vt-flow'
 import {
+  isVtFlowStateChangedEvent,
   SubjectInboundTransport,
   SubjectOutboundTransport,
-  type BaseAgentModules,
+  waitForEvent,
   type SubjectMessage,
   type VsAgent,
 } from '@verana-labs/vs-agent-sdk'
 import { Subject } from 'rxjs'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { isVtFlowStateChangedEvent, startAgent, waitForEvent } from './__mocks__'
+import { startAgent } from './__mocks__'
 
 /**
  * Two-agent in-memory vt-flow round-trip covering thread-id correlation,
