+++ packages/agent-sdk/src/utils/testing/index.ts
@@ -1,2 +1,3 @@
 export * from './SubjectInboundTransport'
 export * from './SubjectOutboundTransport'
+export * from './helpers'
